Fit-applicant freelancer list #108 - 지원한 프리랜서 목록 '확인하기' 버튼 클릭 시 전체 모달 동시 오픈 되고 있어 컴포넌트 분리
